[#468] 셀 수정
